+++ pkg/subscription/list.go
@@ -6,6 +6,15 @@ import (
 	profilesv1 "github.com/weaveworks/profiles/api/v1alpha1"
 )
 
+// SubscriptionSummary contains a summary of a subscription
+type SubscriptionSummary struct {
+	Name      string
+	Namespace string
+	Version   string
+	Profile   string
+	Catalog   string
+}
+
 // List returns a list of subscriptions
 func (sm *Manager) List() ([]SubscriptionSummary, error) {
 	var subscriptions profilesv1.ProfileSubscriptionList
